+++ Client/src/components/AddRecipe/ParameterDroppable.tsx
@@ -1,0 +1,19 @@
+import { useDroppable } from "@dnd-kit/core";
+import { ReactNode } from "react";
+
+const ParameterDroppable = ({ children }: { children: ReactNode }) => {
+  const { setNodeRef } = useDroppable({
+    id: "droppable",
+  });
+
+  return (
+    <div
+      ref={setNodeRef}
+      className="h-full w-full flex flex-col gap-1 col-span-6"
+    >
+      {children}
+    </div>
+  );
+};
+
+export default ParameterDroppable;
