+++ Client/src/pages/AddRecipe.tsx
@@ -1,34 +1,19 @@
-import { Tab, TabGroup, TabList, TabPanel, TabPanels } from "@headlessui/react";
-import { CheckIcon } from "@heroicons/react/20/solid";
+import { TabGroup, TabPanel, TabPanels } from "@headlessui/react";
+import useGetParameters from "@hooks/useGetParameters";
 import {
-  ArrowPathIcon,
-  ChevronDownIcon,
-  ChevronUpIcon,
-  DocumentDuplicateIcon,
-  PlusIcon,
-  XMarkIcon,
-} from "@heroicons/react/24/outline";
-import useGetParameters, {
-  AllParameters,
-  DecimalParameter,
-  TextParameter,
-} from "@hooks/useGetParameters";
-import { IntegerParameter } from "@hooks/useGetParameters";
-import { ReactNode } from "react";
-import { useDroppable, useDraggable } from "@dnd-kit/core";
-import {
-  defaultAnimateLayoutChanges,
   horizontalListSortingStrategy,
   SortableContext,
-  useSortable,
 } from "@dnd-kit/sortable";
-import clsx from "clsx";
 import { RecipeDragProvider } from "@hooks/useRecipeDragContext";
 import {
   AddRecipeProvider,
   useAddRecipeContext,
 } from "@hooks/useAddRecipeContext";
-import DragHandle from "@components/Shared/DragHandle";
+import AddRecipeActions from "@components/AddRecipe/AddRecipeActions";
+import ParameterBox from "@components/AddRecipe/ParameterBox";
+import ParameterSelectList from "@components/AddRecipe/ParameterSelectList";
+import ParameterDroppable from "@components/AddRecipe/ParameterDroppable";
+import StepTabList from "@components/AddRecipe/StepTabList";
 
 const AddRecipeWithContexts = () => {
   const { data: parameters, isLoading, isError } = useGetParameters();
@@ -108,237 +93,6 @@ const AddRecipe = () => {
               </TabPanels>
             </div>
           </TabGroup>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const AddRecipeActions = () => {
-  const { updateRecipe } = useAddRecipeContext();
-  return (
-    <div className="flex border gap-2 border-gray-200 rounded-md bg-white overflow-auto p-2 justify-center shadow-sm">
-      <button className="p-2 border border-gray-200 rounded hover:bg-gray-100 hover:border-red-300 transition-colors duration-100">
-        <ArrowPathIcon
-          onClick={() => updateRecipe({ id: "", name: "", steps: [] })}
-          className="h-5 w-5"
-        />
-      </button>
-      <button className="p-2 border border-gray-200 rounded hover:bg-gray-50 hover:border-green-300 transition-colors duration-100 group">
-        <CheckIcon className="h-5 w-5" />
-      </button>
-    </div>
-  );
-};
-
-type ParameterSelectListProps = {
-  parameters: AllParameters[];
-};
-
-const ParameterSelectList = ({ parameters }: ParameterSelectListProps) => {
-  return (
-    <div className="flex flex-col flex-grow border border-gray-200 rounded-md bg-white shadow-sm overflow-auto">
-      <div className="p-4 border-b border-gray-200 rounded-t-md">
-        <p>Parameters</p>
-      </div>
-      <div className="flex flex-col gap-2 py-2">
-        {parameters.map((parameter) => (
-          <ParameterSelect key={parameter.id} parameter={parameter} />
-        ))}
-      </div>
-    </div>
-  );
-};
-
-const StepTabList = () => {
-  const { addStep, recipe } = useAddRecipeContext();
-  return (
-    <TabList className="flex gap-1 w-full p-2 overflow-x-auto">
-      <div className="overflow-x-auto flex gap-1 w-full">
-        {recipe.steps.map((_, index) => (
-          <StepTab key={index} index={index} />
-        ))}
-      </div>
-      <button
-        onClick={() => addStep()}
-        className="p-2 rounded border border-gray-200 bg-white flex items-center justify-center aspect-square hover:bg-gray-100 z-50"
-      >
-        <PlusIcon className="h-5 aspect-square" />
-      </button>
-    </TabList>
-  );
-};
-
-type StepTabProps = {
-  index: number;
-};
-
-const StepTab = ({ index }: StepTabProps) => {
-  const { copyStepAsNext, removeStep } = useAddRecipeContext();
-  return (
-    <Tab
-      as="div"
-      key={index}
-      className="w-full p-2 rounded border border-gray-200 bg-white relative group data-[selected]:bg-gray-50 focus:outline-none"
-    >
-      <p className="text-sm whitespace-nowrap text-center">Step {index + 1}</p>
-      <div className="absolute right-2 top-1/2 -translate-y-1/2">
-        <div className="flex items-center justify-center">
-          <button
-            onClick={() => copyStepAsNext(index)}
-            className="group-hover:block rounded p-px hidden hover:bg-gray-100 bg-white"
-          >
-            <DocumentDuplicateIcon className="h-5 w-5" />
-          </button>
-          <button
-            onClick={() => removeStep(index)}
-            className="group-hover:block rounded p-px hidden hover:bg-gray-100 bg-white"
-          >
-            <XMarkIcon className="h-5 w-5" />
-          </button>
-        </div>
-      </div>
-    </Tab>
-  );
-};
-
-const ParameterDroppable = ({ children }: { children: ReactNode }) => {
-  const { isOver, setNodeRef } = useDroppable({
-    id: "droppable",
-  });
-  const style = {
-    color: isOver ? "green" : undefined,
-  };
-
-  return (
-    <div
-      ref={setNodeRef}
-      style={style}
-      className="h-full w-full flex flex-col gap-1 col-span-6"
-    >
-      {children}
-    </div>
-  );
-};
-
-type ParameterBoxProps = {
-  parameter: IntegerParameter | DecimalParameter | TextParameter;
-};
-
-export const ParameterSelect = ({ parameter }: ParameterBoxProps) => {
-  const { attributes, listeners, setNodeRef, isDragging } = useDraggable({
-    id: parameter.name,
-    data: parameter,
-  });
-
-  return (
-    !isDragging && (
-      <div className="px-2" ref={setNodeRef}>
-        <div className="border border-gray-200 rounded flex items-center bg-white justify-between shadow-sm">
-          <div className="p-2">
-            <p className="text-xs">{parameter.name}</p>
-          </div>
-          <div className="bg-gray-100 border-s border-gray-200 p-1">
-            <DragHandle listeners={listeners} attributes={attributes} />
-          </div>
-        </div>
-      </div>
-    )
-  );
-};
-
-function animateLayoutChanges(args) {
-  const { isSorting, wasDragging } = args;
-
-  if (isSorting || wasDragging) {
-    return defaultAnimateLayoutChanges(args);
-  }
-
-  return true;
-}
-
-const ParameterBox = ({ parameter }: ParameterBoxProps) => {
-  const {
-    removeParameter,
-    moveParameterUp,
-    moveParameterDown,
-    currentStep,
-    updateParameter,
-  } = useAddRecipeContext();
-  const {
-    listeners,
-    attributes,
-    setNodeRef,
-    transform,
-    transition,
-    isDragging,
-  } = useSortable({
-    animateLayoutChanges,
-    id: parameter.id,
-    data: parameter,
-  });
-
-  return (
-    <div
-      id={parameter.id}
-      ref={setNodeRef}
-      style={{
-        transform: transform
-          ? `translate3d(0, ${transform.y}px, 0)`
-          : undefined,
-        transition,
-      }}
-      className={clsx(
-        "rounded-md border border-gray-200 bg-white shadow-sm !tanslate-x-0",
-        isDragging && "z-50",
-      )}
-    >
-      <div className="border-b border-gray-200 rounded-t-md bg-gray-100 flex justify-between">
-        <p className="p-2 text-sm">{parameter.name}</p>
-        <div className="flex gap-2 px-2 items-center justify-center">
-          <button onClick={() => moveParameterUp(currentStep, parameter.id)}>
-            <ChevronUpIcon className="h-4 w-4" />
-          </button>
-          <button onClick={() => moveParameterDown(currentStep, parameter.id)}>
-            <ChevronDownIcon className="h-4 w-4" />
-          </button>
-          <button onClick={() => removeParameter(currentStep, parameter.id)}>
-            <XMarkIcon className="h-4 w-4" />
-          </button>
-        </div>
-      </div>
-      <div className="p-2">
-        <div className="flex flex-col gap-2">
-          <div className="flex items-center justify-start rounded-md border border-gray-200">
-            <p className="text-xs border-e border-gray-200 p-2 bg-gray-100 text-gray-700">
-              default
-            </p>
-            <input
-              className="rounded-md w-full text-sm ms-2 focus:outline-none"
-              type="text"
-              defaultValue={parameter.defaultValue}
-              value={parameter.value}
-              onChange={(val) => {
-                const newParameter =
-                  parameter.$type === "text"
-                    ? { ...parameter, value: val.currentTarget.value }
-                    : {
-                        ...parameter,
-                        value: parseInt(val.currentTarget.value),
-                      };
-                updateParameter(currentStep, newParameter);
-              }}
-            />
-            <DragHandle attributes={attributes} listeners={listeners} />
-          </div>
-          {(parameter.$type === "integer" || parameter.$type === "decimal") && (
-            <div className="flex items-center justify-start rounded-md border border-gray-200">
-              <p className="text-xs border-e border-gray-200 p-2 bg-gray-100 text-gray-700">
-                unit
-              </p>
-              <p className="text-xs px-2">{parameter.unit}</p>
-            </div>
-          )}
         </div>
       </div>
     </div>
